requirements.txt
Flask==2.1.0
line-bot-sdk==1.19.0
python-dotenv==0.19.1
Pillow==8.4.0
openai==0.11.3
clip-interrogator==0.1.7
